docker-mysql/backend/requirements.txt
# --- Core Framework ---
django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1

# --- Database ---
mysqlclient>=2.2.4
# 如果在 Windows 上安装 mysqlclient 出错，可改用：
# PyMySQL>=1.1.0

# --- Cache / Redis ---
django-redis>=5.4.0
redis>=5.0.0

# --- CORS ---
django-cors-headers>=4.3.1

# --- Environment Variables ---
python-dotenv>=1.0.1

# --- Optional: for async tasks / Celery ---
# celery>=5.3.6

# --- Utilities ---
requests>=2.31.0
